+++ package/com.unity.formats.usd/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changes in usd-unitysdk for Unity
 
+## [1.0.2-preview.1] - 2019-11-06
+### Added
+- Option to handle basis conversion "as FBX" (#129)
+- Support for "Sphere" schema (#128)
+
+### Changed
+- Fix the "Decompose" function to handle matrices with very small determinant (#130, #142)
+- Allow exporting transparency from Standard shader (#131)
+- Fix the "ImportMesh" sample to works in edit mode (#128)
+- Handle skinned primitive that inherits skeleton data (joint indices and weight) from parent prim but don't define those themselves (#128)
+
+
 ## [1.0.1-preview.1] - 2019-06-30
 ### Changes